refactor(archtest): Governance/Contract 规则族迁 module-path-agnostic CellRule [#1638]
M3 PR-8a：把 6 个 Governance/Contract 族 archtest 规则文件迁成 module-path-agnostic，
按 #1084 PR-1 配方（#1632 / #1635 同系列先例）。每条规则检测逻辑从 _test.go MOVE 进
可被外部仓库 import 的非 test Check*(t, ConfigForExternalCell) []Diagnostic；平台符号
路径从 external.go::PlatformModulePath 派生（消除 bare github.com/ghbvf/gocell 字面量）；
瘦身 _test.go dogfood 同一 Check*（单源）并保留全部 RED/负例/blind-spot 自检。

迁移规则（register=none，external.go 未改）：
- CONTRACTTEST-LOADBYID-LITERAL-01            (contracttest_loadbyid_literal)
- PARSER-MATCHER-EXAMPLES-SYMMETRY-01         (parser_matcher_examples_symmetry)
- CONTRACT-PATH-QUERY-COVERAGE-01 / -PARAM-NAME-LITERAL-01 (contract_path_query_coverage)
- SERVICEOWNED-HANDLER-OWNER-CHECK-01         (serviceowned_handler_owner_check)
- GOVERNANCE-RULES-REGISTRATION-GUARD-01 / -CODE-CONST-SINGLE-SOURCE-01 /
  -ERROR-FIX-FIELD-01 / -CODE-DETECT-BINDING-01 / -EMITTER-CONSTRUCTORS-NEVER-FUNCVALUE-01 (governance_rules_invariants)
- IMPL-DECL-COVER-01 / HANDLER-DECL-COVER-01 / EMIT-DECL-COVER-01 /
  DEAD-CONTRACT-01 / DEAD-CODE-01            (reverse_coverage_invariants)

behavior 零变 + AI-robust 评级零变（dogfood + RED 自检证明 scanner 未变）。
IMPL-DECL-COVER-01 为可移植 cell-architecture 规则，本批 register=none 保 behavior 零变，
StandardCellRules 注册由 epic #1302 holistically 处理。

ref: aec79118e (#1632) / 4f3496e70 (#1635) 同配方
