+++ tools/archtest/reverse_coverage_invariants_test.go
@@ -12,6 +12,10 @@
 // contract.yaml. AI-robust grading per rule documented in each Test*
 // godoc.
 //
+// Each main Test* dogfoods the Check* function from reverse_coverage_invariants.go.
+// Negative self-checks (_Detects* / FloorScan / ProviderEndpointMirrorsMetadata)
+// exercise the scanner internals directly to confirm detection paths are live.
+//
 // ref: TNG/ArchUnit archunit/src/main/java/com/tngtech/archunit/library/Architectures.java@main
 // ref: TNG/ArchUnit archunit/src/main/java/com/tngtech/archunit/core/domain/JavaClass.java@main
 // ref: kubernetes/kubernetes hack/verify-imports.sh@master
@@ -21,323 +25,11 @@ import (
 	"go/ast"
 	"go/parser"
 	"go/token"
-	"go/types"
-	"os"
-	"path/filepath"
 	"strings"
-	"sync"
 	"testing"
 
-	"gopkg.in/yaml.v3"
-
 	"github.com/ghbvf/gocell/kernel/metadata"
-	"github.com/ghbvf/gocell/tools/archtest/internal/typeseval"
-	"github.com/ghbvf/gocell/tools/typesutil"
 )
-
-// ---------------------------------------------------------------------------
-// Shared contract loader (lazy sync.Once, used by HANDLER, EMIT, DEAD rules)
-// ---------------------------------------------------------------------------
-
-// contractDoc is the minimal shape of contract.yaml needed for the 5 rules.
-type contractDoc struct {
-	ID        string            `yaml:"id"`
-	Kind      string            `yaml:"kind"`
-	Lifecycle string            `yaml:"lifecycle"`
-	OwnerCell string            `yaml:"ownerCell"`
-	Endpoints contractEndpoints `yaml:"endpoints"`
-	Triggers  []string          `yaml:"triggers"`
-	FilePath  string            `yaml:"-"` // absolute path to the contract.yaml
-}
-
-// contractEndpoints mirrors kernel/metadata.EndpointsMeta: actorSubscribers
-// lives under endpoints (not at the top level). Confirmed by all 4 contracts
-// using this field today (event.audit.appended.v1 + 3 example events).
-//
-// The per-kind provider field differs (server/publisher/handler/provider); the
-// DEAD-CONTRACT-01 provider check must read the SAME field metadata's
-// ContractMeta.ProviderEndpoint reads for that kind — see contractProviderEndpoint.
-type contractEndpoints struct {
-	Server           string   `yaml:"server"`
-	Publisher        string   `yaml:"publisher"`
-	Handler          string   `yaml:"handler"`  // kind:command provider (mirrors metadata)
-	Provider         string   `yaml:"provider"` // kind:projection provider (mirrors metadata)
-	Invokers         []string `yaml:"invokers"`
-	ActorSubscribers []string `yaml:"actorSubscribers"`
-}
-
-// contractProviderEndpoint returns the provider endpoint for a contract, reading
-// the per-kind provider field exactly as kernel/metadata.ContractMeta.ProviderEndpoint
-// does (http/grpc/saga→server, event→publisher, command→handler,
-// projection→provider, webhook→ownerCell). DEAD-CONTRACT-01 uses this instead of a
-// hardcoded endpoints.server fallback so a valid contract declaring only its
-// kind-correct provider (e.g. a kind:command with endpoints.handler and no ownerCell)
-// is not false-flagged. Parity with metadata is locked by
-// TestDeadContractCover_ProviderEndpointMirrorsMetadata — if metadata's per-kind
-// provider semantics change, that test fails until this mirror is updated.
-func contractProviderEndpoint(c contractDoc) string {
-	switch c.Kind {
-	case "http", "grpc", "saga":
-		return c.Endpoints.Server
-	case "event":
-		return c.Endpoints.Publisher
-	case "command":
-		return c.Endpoints.Handler
-	case "projection":
-		return c.Endpoints.Provider
-	case "webhook":
-		return c.OwnerCell
-	default:
-		return ""
-	}
-}
-
-// contractProviderFieldLabel names the provider endpoint field for a kind, for
-// human-readable DEAD-CONTRACT-01 diagnostics (keeps the message pointing at the
-// exact field a fix should populate).
-func contractProviderFieldLabel(kind string) string {
-	switch kind {
-	case "command":
-		return "endpoints.handler"
-	case "projection":
-		return "endpoints.provider"
-	case "event":
-		return "endpoints.publisher"
-	default:
-		return "endpoints.server"
-	}
-}
-
-var (
-	contractsOnce sync.Once
-	contractsAll  []contractDoc
-	contractsErr  error
-)
-
-// loadReverseCoverageContracts loads all contract.yaml files from the contracts/ tree
-// and examples/*/contracts/ trees. Results are cached via sync.Once across the test binary lifetime.
-func loadReverseCoverageContracts(t *testing.T) []contractDoc {
-	t.Helper()
-	root := findModuleRoot(t)
-	contractsOnce.Do(func() {
-		contractsAll, contractsErr = loadContractDocs(root)
-	})
-	if contractsErr != nil {
-		t.Fatalf("[%s] loadReverseCoverageContracts: %v", t.Name(), contractsErr)
-	}
-	return contractsAll
-}
-
-// loadContractDocs scans contracts/ and examples/*/contracts/ for every
-// contract.yaml and parses each one. Uses loadContentFiles (the archtest
-// framework) to avoid forbidden os.ReadDir / filepath.WalkDir calls.
-func loadContractDocs(root string) ([]contractDoc, error) {
-	// Scan under two top-level dirs: contracts/ (platform) and examples/ (examples).
-	// MatchRels filter restricts to files named "contract.yaml"; the path for
-	// examples must also contain a "/contracts/" segment to avoid picking up
-	// non-contract YAML files scattered under examples/.
-	scope := DirsScope(
-		root, []string{"contracts", "examples"},
-		MatchRels(func(rel string) bool {
-			if !strings.HasSuffix(rel, "/contract.yaml") {
-				return false
-			}
-			// For examples/ paths, require a /contracts/ segment to avoid
-			// picking up non-contract YAML files.
-			if strings.HasPrefix(rel, "examples/") {
-				return strings.Contains(rel, "/contracts/")
-			}
-			return true
-		}),
-	)
-
-	files, loadErr := loadContentFiles(scope, []string{".yaml"})
-	if loadErr != nil {
-		return nil, loadErr
-	}
-
-	docs := make([]contractDoc, 0, len(files))
-	for _, fc := range files {
-		var doc contractDoc
-		if parseErr := yaml.Unmarshal(fc.Bytes, &doc); parseErr != nil {
-			return nil, parseErr
-		}
-		doc.FilePath = fc.AbsPath
-		docs = append(docs, doc)
-	}
-	return docs, nil
-}
-
-// extractSourceComment scans the first 10 lines of a generated Go file for a
-// "// source: <rel-path>" comment and returns the relative path (forward-slash
-// form, as written by codegen).
-func extractSourceComment(content []byte) string {
-	for _, line := range strings.SplitN(string(content), "\n", 10) {
-		const prefix = "// source: "
-		if strings.HasPrefix(line, prefix) {
-			return strings.TrimSpace(strings.TrimPrefix(line, prefix))
-		}
-	}
-	return ""
-}
-
-// loadGeneratedSourceMap scans generated/contracts/<kind>/**/<suffix> files
-// and builds a map from module-relative generated package import path to the
-// absolute path of the declaring contract.yaml. The mapping is extracted from
-// the "// source: <rel-path>" comment that codegen writes into each generated
-// file.
-//
-// This handles the codegen "internal" → "internalapi" segment remapping
-// correctly (the comment carries the true source path, not the dot-to-slash
-// derived path).
-func loadGeneratedSourceMap(root, modPath string, kindDir, fileSuffix string) (map[string]string, error) {
-	scope := DirsScope(
-		root, []string{"generated/contracts/" + kindDir},
-		IncludeGenerated(),
-		MatchRels(func(rel string) bool {
-			return strings.HasSuffix(rel, fileSuffix)
-		}),
-	)
-
-	files, loadErr := loadContentFiles(scope, []string{".go"})
-	if loadErr != nil {
-		return nil, loadErr
-	}
-
-	result := make(map[string]string, len(files))
-	for _, fc := range files {
-		// Derive generated package import path from the file's directory.
-		rel := fc.Rel
-		dir := rel[:strings.LastIndex(rel, "/")]
-		genPkgPath := modPath + "/" + dir
-
-		srcRel := extractSourceComment(fc.Bytes)
-		if srcRel == "" {
-			continue
-		}
-		result[genPkgPath] = filepath.Join(root, filepath.FromSlash(srcRel))
-	}
-	return result, nil
-}
-
-// loadGeneratedHTTPServiceMap is the HTTP-Service-interface-specific variant
-// of loadGeneratedSourceMap, used by HANDLER-DECL-COVER-01 and
-// DEAD-CONTRACT-01 to locate the Service interface package for each http
-// contract.
-func loadGeneratedHTTPServiceMap(root, modPath string) (map[string]string, error) {
-	return loadGeneratedSourceMap(root, modPath, "http", "/iface_gen.go")
-}
-
-// loadGeneratedAllSourceMap covers every generated contract package
-// (http/event/projection), used by DEAD-CODE-01 to detect production imports
-// of any deprecated contract's generated package (not just http). All three
-// kinds emit `iface_gen.go` consistently with a `// source: <rel>` header.
-func loadGeneratedAllSourceMap(root, modPath string) (map[string]string, error) {
-	merged := make(map[string]string)
-	for _, kind := range []string{"http", "event", "projection"} {
-		m, err := loadGeneratedSourceMap(root, modPath, kind, "/iface_gen.go")
-		if err != nil {
-			return nil, err
-		}
-		for k, v := range m {
-			merged[k] = v
-		}
-	}
-	return merged, nil
-}
-
-// ---------------------------------------------------------------------------
-// Slice subscriber index (for DEAD-CONTRACT-01 event subscriber detection)
-// ---------------------------------------------------------------------------
-
-// sliceSubscriberEntry describes a slice that subscribes to an event contract.
-type sliceSubscriberEntry struct {
-	BelongsToCell string
-	ContractID    string
-}
-
-// loadSliceSubscribers scans cells/**/slice.yaml AND examples/**/slice.yaml
-// for subscribe contractUsages. Scope must mirror loadContractDocs so that an
-// examples slice subscribing to a platform contract is visible to
-// DEAD-CONTRACT-01.
-func loadSliceSubscribers(root string) ([]sliceSubscriberEntry, error) {
-	scope := DirsScope(
-		root, []string{"cells", "examples"},
-		MatchRels(func(rel string) bool {
-			return strings.HasSuffix(rel, "/slice.yaml")
-		}),
-	)
-
-	files, loadErr := loadContentFiles(scope, []string{".yaml"})
-	if loadErr != nil {
-		return nil, loadErr
-	}
-
-	var entries []sliceSubscriberEntry
-	for _, fc := range files {
-		var doc struct {
-			BelongsToCell  string `yaml:"belongsToCell"`
-			ContractUsages []struct {
-				Contract string `yaml:"contract"`
-				Role     string `yaml:"role"`
-			} `yaml:"contractUsages"`
-		}
-		if parseErr := yaml.Unmarshal(fc.Bytes, &doc); parseErr != nil {
-			return nil, parseErr
-		}
-		for _, cu := range doc.ContractUsages {
-			if cu.Role == "subscribe" && cu.Contract != "" {
-				entries = append(entries, sliceSubscriberEntry{
-					BelongsToCell: doc.BelongsToCell,
-					ContractID:    cu.Contract,
-				})
-			}
-		}
-	}
-	return entries, nil
-}
-
-// loadSliceWebhookWiring scans cells/**/slice.yaml AND examples/**/slice.yaml
-// for webhook-receive / webhook-dispatch contractUsages and returns the set of
-// contract IDs that have at least one such wiring. Mirrors loadSliceSubscribers
-// (the contract.yaml Receivers/Dispatchers fields are derived, yaml:"-", and so
-// are NOT populated by the static loader DEAD-CONTRACT-01 uses; the wiring must
-// be discovered from the slice scan, exactly as event subscribers are).
-func loadSliceWebhookWiring(root string) (map[string]bool, error) {
-	scope := DirsScope(
-		root, []string{"cells", "examples"},
-		MatchRels(func(rel string) bool {
-			return strings.HasSuffix(rel, "/slice.yaml")
-		}),
-	)
-
-	files, loadErr := loadContentFiles(scope, []string{".yaml"})
-	if loadErr != nil {
-		return nil, loadErr
-	}
-
-	wired := make(map[string]bool)
-	for _, fc := range files {
-		var doc struct {
-			ContractUsages []struct {
-				Contract string `yaml:"contract"`
-				Role     string `yaml:"role"`
-			} `yaml:"contractUsages"`
-		}
-		if parseErr := yaml.Unmarshal(fc.Bytes, &doc); parseErr != nil {
-			return nil, parseErr
-		}
-		for _, cu := range doc.ContractUsages {
-			if cu.Contract == "" {
-				continue
-			}
-			if cu.Role == "webhook-receive" || cu.Role == "webhook-dispatch" {
-				wired[cu.Contract] = true
-			}
-		}
-	}
-	return wired, nil
-}
 
 // ---------------------------------------------------------------------------
 // IMPL-DECL-COVER-01
@@ -348,6 +40,10 @@ func loadSliceWebhookWiring(root string) (map[string]bool, error) {
 // different cell cells/<B>/... unless the import path is under
 // cells/<B>/<B>test/ (the public test helper boundary).
 //
+// Mechanism: AST ImportSpec scan under cells/. For each file, ownerCell is
+// extracted from its module-relative path; any import under cells/<B>/ where
+// B ≠ ownerCell and B is not the <B>test/* structural helper boundary is flagged.
+//
 // AI-robust funnel evaluation:
 //   - upstream: Hard — every Go file's ImportSpec must be parsed by go/parser
 //     to be visible to the Go toolchain; there is no escape from ImportSpec
@@ -355,141 +51,12 @@ func loadSliceWebhookWiring(root string) (map[string]bool, error) {
 //   - downstream: Hard — AST ImportSpec.Path.Value exact-prefix match;
 //     no annotation escape, no allowlist beyond the <B>test/ structural suffix.
 //
-// Blind-spot self-check (BlindSpots outside the assertions above):
-//   - Dot imports: `import . "cells/A/…"` yields a bare Ident at use-site, not
-//     a SelectorExpr — the import path string itself is still captured in
-//     ImportSpec.Path.Value and is checked here.
-//   - Alias imports: `import foo "cells/A/…"` — same, import path unchanged.
-//   - Blank imports: `import _ "cells/A/…"` — same, ImportSpec.Path.Value
-//     is present regardless of name; checked by TestImplDeclCover_DetectsMissingImport.
-//   - Generated cell_gen.go / *_gen.go within the SAME cell: allowed (same cell).
-//   - Test files (*_test.go): excluded by scope (no IncludeTests()).
+// Blind-spot self-check (TestImplDeclCover_DetectsMissingImport):
+//   - Qualified import: covered (normal case, tested).
+//   - Blank import: covered (ImportSpec.Path.Value is the same regardless of name).
 func TestImplDeclCover(t *testing.T) {
 	t.Parallel()
-	root := findModuleRoot(t)
-	modPath, err := moduleImportPath(root)
-	if err != nil {
-		t.Fatalf("IMPL-DECL-COVER-01: read module path: %v", err)
-	}
-	cellsPrefix := modPath + "/cells/"
-
-	scope := DirsScope(root, []string{"cells"}, MatchRels(func(rel string) bool {
-		// production only: no _test.go, no testdata
-		return !strings.HasSuffix(rel, "_test.go") && !strings.Contains(rel, "/testdata/")
-	}))
-
-	diags := Run(t, AST(scope), func(p *Pass) []Diagnostic {
-		var d []Diagnostic
-		for _, f := range p.Files {
-			rel := p.Rel(f)
-
-			ownerCell := extractCellName(rel)
-			if ownerCell == "" {
-				continue
-			}
-			for _, imp := range f.Imports {
-				if imp.Path == nil {
-					continue
-				}
-				impPath := strings.Trim(imp.Path.Value, `"`)
-				if !strings.HasPrefix(impPath, cellsPrefix) {
-					continue
-				}
-
-				impCell := extractCellNameFromImport(cellsPrefix, impPath)
-				if impCell == "" || impCell == ownerCell {
-					continue
-				}
-
-				testBoundary := cellsPrefix + impCell + "/" + impCell + "test/"
-				if strings.HasPrefix(impPath, testBoundary) {
-					continue
-				}
-				pos := p.Fset.Position(imp.Pos())
-				d = append(d, Diagnostic{
-					Rel:  rel,
-					Line: pos.Line,
-					Message: "cross-cell import from " + ownerCell + " to cells/" + impCell +
-						" bypasses contract boundary (only cells/<X>/<X>test/* allowed): " + impPath,
-				})
-			}
-		}
-		return d
-	})
-
-	Report(t, "IMPL-DECL-COVER-01", diags)
-}
-
-// extractCellName returns the cell directory name for a file at rel path
-// (module-relative slash path) under cells/. Returns "" when not under cells/.
-// For relative paths (relative to module root), uses the "cells/" prefix directly.
-func extractCellName(rel string) string {
-	// rel is a module-relative slash path like "cells/accesscore/slices/foo/bar.go"
-	const pfx = "cells/"
-	if !strings.HasPrefix(rel, pfx) {
-		return ""
-	}
-	rest := rel[len(pfx):]
-	idx := strings.Index(rest, "/")
-	if idx < 0 {
-		return rest
-	}
-	return rest[:idx]
-}
-
-// extractCellIDFromPkgPath returns the cell ID for a Go package path, treating
-// any "/cells/<X>/..." segment as contract-owned. Handles both platform cells
-// (modPath+"/cells/<X>/...") and example cells
-// (modPath+"/examples/<demo>/cells/<X>/...") by structural marker, not by
-// hand-maintained allowlist. Returns "" for:
-//   - non-cell packages (runtime/, adapters/, kernel/, examples/<demo>/[non-cells])
-//   - the cell's public test-helper package "cells/<X>/<X>test/..." (structural
-//     test boundary — fixtures construct outbox.Entry literals for consumer
-//     tests, those are not real emissions; mirrors the IMPL-DECL-COVER-01
-//     <X>test exception)
-//
-// Both exclusions are structural (path-shape), not allowlist.
-func extractCellIDFromPkgPath(modPath, pkgPath string) string {
-	if !strings.HasPrefix(pkgPath, modPath+"/") {
-		return ""
-	}
-	const marker = "/cells/"
-	idx := strings.Index(pkgPath, marker)
-	if idx < 0 {
-		return ""
-	}
-	tail := pkgPath[idx+len(marker):]
-	cellID := tail
-	if next := strings.Index(tail, "/"); next >= 0 {
-		cellID = tail[:next]
-		// Structural test-helper boundary: cells/<X>/<X>test/...
-		rest := tail[next+1:]
-		nextSeg := rest
-		if slashIdx := strings.Index(rest, "/"); slashIdx >= 0 {
-			nextSeg = rest[:slashIdx]
-		}
-		if nextSeg == cellID+"test" {
-			return ""
-		}
-	}
-	return cellID
-}
-
-// extractCellNameFromImport extracts the cell name from a full import path like
-// "github.com/ghbvf/gocell/cells/accesscore/..." for the platform-only
-// IMPL-DECL-COVER-01 cross-cell ban (which scans cells/ exclusively).
-// EMIT-DECL-COVER-01 and any rule scanning examples/ must use
-// extractCellIDFromPkgPath instead.
-func extractCellNameFromImport(cellsImportPrefix, impPath string) string {
-	if !strings.HasPrefix(impPath, cellsImportPrefix) {
-		return ""
-	}
-	rest := impPath[len(cellsImportPrefix):]
-	idx := strings.Index(rest, "/")
-	if idx < 0 {
-		return rest
-	}
-	return rest[:idx]
+	Report(t, "IMPL-DECL-COVER-01", CheckImplDeclCover(t, ConfigForExternalCell{}))
 }
 
 // TestImplDeclCover_DetectsMissingImport is the negative-fixture self-check for
@@ -506,7 +73,7 @@ func extractCellNameFromImport(cellsImportPrefix, impPath string) string {
 // confirm the AST scanner path (not just the string-utility path) works.
 func TestImplDeclCover_DetectsMissingImport(t *testing.T) {
 	t.Parallel()
-	const modPath = "github.com/ghbvf/gocell"
+	const modPath = PlatformModulePath
 	cellsPrefix := modPath + "/cells/"
 
 	// 1. String-utility path.
@@ -612,173 +179,7 @@ import (
 //   - Satellite-module invisibility: caught by the sawSatelliteExamplePkg sentinel.
 func TestHandlerDeclCover(t *testing.T) {
 	t.Parallel()
-	root := findModuleRoot(t)
-	modPath, err := moduleImportPath(root)
-	if err != nil {
-		t.Fatalf("HANDLER-DECL-COVER-01: read module path: %v", err)
-	}
-
-	// Build genPkgPath → contractYamlAbsPath from the "// source:" comments in
-	// generated/contracts/http/*/iface_gen.go. This is the correct mapping
-	// because codegen remaps some source path segments (e.g. "internal" →
-	// "internalapi") that the dot-to-slash ID transform cannot reproduce.
-	genHTTPSourceMap, mapErr := loadGeneratedHTTPServiceMap(root, modPath)
-	if mapErr != nil {
-		t.Fatalf("HANDLER-DECL-COVER-01: loadGeneratedHTTPServiceMap: %v", mapErr)
-	}
-
-	// activeHTTPContracts: genPkgPath → true when the source contract.yaml
-	// exists and is lifecycle: active. Uses the source map to look up the
-	// contract's lifecycle from the YAML rather than from contractDoc (which
-	// could be stale if loadContractDocs misses examples/ or internalapi).
-	activeHTTPContracts := make(map[string]bool)
-	for genPkg, contractPath := range genHTTPSourceMap {
-		raw, readErr := os.ReadFile(filepath.Clean(contractPath))
-		if readErr != nil {
-			continue // missing on disk → not active
-		}
-		var doc struct {
-			Lifecycle string `yaml:"lifecycle"`
-		}
-		if yaml.Unmarshal(raw, &doc) == nil && doc.Lifecycle == "active" {
-			activeHTTPContracts[genPkg] = true
-		}
-	}
-
-	generatedHTTPPrefix := modPath + "/generated/contracts/http/"
-	cellsPrefix := modPath + "/cells/"
-	examplesPrefix := modPath + "/examples/"
-
-	type ifaceEntry struct {
-		pkgPath  string
-		ifaceTyp *types.Interface
-		pos      token.Position // source position of the Service type declaration
-	}
-	type implEntry struct {
-		named   *types.Named
-		pkgPath string
-		name    string
-		pos     token.Position // source position of the concrete type declaration
-	}
-
-	var genServiceIfaces []ifaceEntry
-	var cellImplTypes []implEntry
-	// Anti-vacuity: prove the workspace load actually reached a satellite example
-	// module (examples/* beyond the in-root demo). If the ModeWorkspace loader
-	// ever regresses to root-only, this stays false and the test fails loudly
-	// rather than silently passing with every satellite HTTP Service impl invisible.
-	demoExamplePrefix := examplesPrefix + "demo/"
-	sawSatelliteExamplePkg := false
-
-	// Build the iface + impl type sets in ONE ModeWorkspace type universe via
-	// LoadProductionPackages — NOT the Typed/ModeModule loader. ModeWorkspace is
-	// mandatory: examples/* are go.work satellite modules (#1556), and a GOWORK=off
-	// load silently matches only the in-root examples/demo, dropping every satellite
-	// example impl so its HTTP Service contracts look unimplemented (false orphan).
-	// .All() (NOT .Production()) keeps the generated/contracts/http packages whose
-	// Service interfaces drive the types.Implements match; iface + impl land in one
-	// type-check universe so ImplementsInterface compares identical packages. Mirrors
-	// DEAD-CONTRACT-01 below + the loadModule funnel in archtest_test.go.
-	modules := findWorkspaceModules(t, root)
-	resolver, lpErr := typeseval.LoadProductionPackages(root, modules, false /* tests */, nil)
-	if lpErr != nil {
-		t.Fatalf("HANDLER-DECL-COVER-01: LoadProductionPackages: %v", lpErr)
-	}
-	for _, pkg := range resolver.All() {
-		if pkg == nil || pkg.Types == nil {
-			continue
-		}
-		pkgPath := pkg.PkgPath
-
-		if strings.HasPrefix(pkgPath, generatedHTTPPrefix) {
-			obj := pkg.Types.Scope().Lookup("Service")
-			if obj == nil {
-				continue
-			}
-			tn, ok := obj.(*types.TypeName)
-			if !ok {
-				continue
-			}
-			named, ok := tn.Type().(*types.Named)
-			if !ok {
-				continue
-			}
-			iface, ok := named.Underlying().(*types.Interface)
-			if !ok {
-				continue
-			}
-			pos := pkg.Fset.Position(tn.Pos())
-			genServiceIfaces = append(genServiceIfaces, ifaceEntry{
-				pkgPath:  pkgPath,
-				ifaceTyp: iface.Complete(),
-				pos:      pos,
-			})
-			continue
-		}
-
-		isCells := strings.HasPrefix(pkgPath, cellsPrefix)
-		isExamples := strings.HasPrefix(pkgPath, examplesPrefix)
-		if !isCells && !isExamples {
-			continue
-		}
-		if isExamples && !strings.HasPrefix(pkgPath, demoExamplePrefix) {
-			sawSatelliteExamplePkg = true
-		}
-
-		pkgScope := pkg.Types.Scope()
-		for _, name := range pkgScope.Names() {
-			obj := pkgScope.Lookup(name)
-			tn, ok := obj.(*types.TypeName)
-			if !ok {
-				continue
-			}
-			named, ok := tn.Type().(*types.Named)
-			if !ok {
-				continue
-			}
-			pos := pkg.Fset.Position(tn.Pos())
-			cellImplTypes = append(cellImplTypes, implEntry{
-				named:   named,
-				pkgPath: pkgPath,
-				name:    name,
-				pos:     pos,
-			})
-		}
-	}
-
-	if len(genServiceIfaces) == 0 {
-		t.Fatal("HANDLER-DECL-COVER-01: no generated http Service interfaces found — scanner may be broken")
-	}
-	if !sawSatelliteExamplePkg {
-		t.Fatal("HANDLER-DECL-COVER-01: workspace load surfaced no satellite example package " +
-			"(examples/* beyond demo) — the ModeWorkspace loader regressed to root-only and satellite " +
-			"HTTP Service impls would be invisible to this orphan-impl cross-check (#1556)")
-	}
-
-	// Cross-check: every impl that satisfies a Service iface must have an active contract.
-	var diags []Diagnostic
-	for _, impl := range cellImplTypes {
-		for _, iface := range genServiceIfaces {
-			if !typesutil.ImplementsInterface(impl.named, iface.ifaceTyp) {
-				continue
-			}
-			// Found impl → iface. Check the contract.yaml exists and is active.
-			if !activeHTTPContracts[iface.pkgPath] {
-				relPath, relErr := filepath.Rel(root, impl.pos.Filename)
-				if relErr != nil || relPath == "" {
-					relPath = impl.pkgPath // fallback for synthetic positions
-				}
-				diags = append(diags, Diagnostic{
-					Rel:  relPath,
-					Line: impl.pos.Line,
-					Message: "type " + impl.name + " in " + impl.pkgPath + " implements " +
-						iface.pkgPath + ".Service but no active contract.yaml " +
-						"found for this generated package",
-				})
-			}
-		}
-	}
-	Report(t, "HANDLER-DECL-COVER-01", diags)
+	Report(t, "HANDLER-DECL-COVER-01", CheckHandlerDeclCover(t, ConfigForExternalCell{}))
 }
 
 // TestHandlerDeclCover_DetectsOrphanImpl is the negative self-check for
@@ -795,7 +196,7 @@ func TestHandlerDeclCover_DetectsOrphanImpl(t *testing.T) {
 	t.Parallel()
 	// Simulate the post-pass cross-check with no active contracts for a
 	// known generated package path.
-	const fakeGenPkg = "github.com/ghbvf/gocell/generated/contracts/http/fake/v1"
+	const fakeGenPkg = PlatformModulePath + "/generated/contracts/http/fake/v1"
 
 	activeHTTPContracts := map[string]bool{
 		// fakeGenPkg is intentionally absent → orphan impl
@@ -863,137 +264,7 @@ func TestHandlerDeclCover_DetectsOrphanImpl(t *testing.T) {
 //     TestEmitDeclCover_DetectsNonConstTopic (documented blind-spot if not).
 func TestEmitDeclCover(t *testing.T) {
 	t.Parallel()
-	root := findModuleRoot(t)
-	modPath, err := moduleImportPath(root)
-	if err != nil {
-		t.Fatalf("EMIT-DECL-COVER-01: read module path: %v", err)
-	}
-
-	contracts := loadReverseCoverageContracts(t)
-	const outboxPkg = "github.com/ghbvf/gocell/kernel/outbox"
-
-	// Build allowed-topic set per cell.
-	// cellID → set of allowed topic strings
-	allowedTopics := make(map[string]map[string]bool)
-	for _, c := range contracts {
-		if c.Lifecycle != "active" {
-			continue
-		}
-		// (a) kind: event with publisher → publisher cell may emit with topic = contract.id
-		if c.Kind == "event" && c.Endpoints.Publisher != "" {
-			cell := c.Endpoints.Publisher
-			if allowedTopics[cell] == nil {
-				allowedTopics[cell] = make(map[string]bool)
-			}
-			allowedTopics[cell][c.ID] = true
-		}
-		// (b) triggers: any contract with ownerCell or endpoints.server lists triggers
-		ownerCell := c.OwnerCell
-		if ownerCell == "" {
-			ownerCell = c.Endpoints.Server
-		}
-		if ownerCell == "" {
-			continue
-		}
-		for _, trigger := range c.Triggers {
-			if allowedTopics[ownerCell] == nil {
-				allowedTopics[ownerCell] = make(map[string]bool)
-			}
-			allowedTopics[ownerCell][trigger] = true
-		}
-	}
-
-	diags := Run(t, Production(TypedOpts{Tests: false}), func(p *Pass) []Diagnostic {
-		if p.Pkg == nil || p.TypesInfo == nil {
-			return nil
-		}
-
-		cellID := extractCellIDFromPkgPath(modPath, p.Pkg.Path())
-		if cellID == "" {
-			return nil
-		}
-
-		var d []Diagnostic
-		for _, f := range p.Files {
-			rel := p.Rel(f)
-
-			EachInSubtree[ast.CallExpr](f, func(call *ast.CallExpr) {
-				fun := call.Fun
-				if idx, ok := fun.(*ast.IndexExpr); ok {
-					fun = idx.X
-				} else if idxl, ok := fun.(*ast.IndexListExpr); ok {
-					fun = idxl.X
-				}
-				pkgPath, name, ok := ResolvePackageRef(p.TypesInfo, fun)
-				if !ok || pkgPath != outboxPkg || name != "Emit" {
-					return
-				}
-
-				if len(call.Args) < 4 {
-					return
-				}
-				topic, isConst := EvaluateConstString(p.TypesInfo, call.Args[3])
-				if !isConst {
-					pos := p.Fset.Position(call.Pos())
-					d = append(d, Diagnostic{
-						Rel:  rel,
-						Line: pos.Line,
-						Message: "non-const topic in outbox.Emit; convert caller to use a const string " +
-							"(EMIT-DECL-COVER-01: non-const topics cannot be statically validated)",
-					})
-					return
-				}
-				if !allowedTopics[cellID][topic] {
-					pos := p.Fset.Position(call.Pos())
-					d = append(d, Diagnostic{
-						Rel:  rel,
-						Line: pos.Line,
-						Message: "outbox.Emit topic " + topic + " not declared in any active contract " +
-							"for cell " + cellID + " (must be an event contract publisher or a triggers entry)",
-					})
-				}
-			})
-
-			EachInSubtree[ast.CompositeLit](f, func(lit *ast.CompositeLit) {
-				typPkg, typName, ok := ResolvePackageRef(p.TypesInfo, lit.Type)
-				if !ok || typPkg != outboxPkg || typName != "Entry" {
-					return
-				}
-
-				kv, found := FindFirstChild[ast.KeyValueExpr](lit, func(kv *ast.KeyValueExpr) bool {
-					keyIdent, ok := kv.Key.(*ast.Ident)
-					return ok && keyIdent.Name == "EventType"
-				})
-				if !found {
-					return
-				}
-				topic, isConst := EvaluateConstString(p.TypesInfo, kv.Value)
-				if !isConst {
-					pos := p.Fset.Position(kv.Pos())
-					d = append(d, Diagnostic{
-						Rel:  rel,
-						Line: pos.Line,
-						Message: "non-const EventType in outbox.Entry literal; convert to const string " +
-							"(EMIT-DECL-COVER-01: non-const topics cannot be statically validated)",
-					})
-					return
-				}
-				if !allowedTopics[cellID][topic] {
-					pos := p.Fset.Position(kv.Pos())
-					d = append(d, Diagnostic{
-						Rel:  rel,
-						Line: pos.Line,
-						Message: "outbox.Entry{EventType: " + topic + "} topic not declared in any " +
-							"active contract for cell " + cellID + " (must be an event contract " +
-							"publisher or a triggers entry)",
-					})
-				}
-			})
-		}
-		return d
-	})
-
-	Report(t, "EMIT-DECL-COVER-01", diags)
+	Report(t, "EMIT-DECL-COVER-01", CheckEmitDeclCover(t, ConfigForExternalCell{}))
 }
 
 // TestEmitDeclCover_DetectsNonConstTopic is the negative self-check for
@@ -1128,338 +399,7 @@ func outboxEmit(ctx context.Context, e interface{}, topic string, p interface{})
 //     additionally locks the diagnostic-emission decision in isolation.
 func TestDeadContractCover(t *testing.T) {
 	t.Parallel()
-	root := findModuleRoot(t)
-	modPath, err := moduleImportPath(root)
-	if err != nil {
-		t.Fatalf("DEAD-CONTRACT-01: read module path: %v", err)
-	}
-
-	contracts := loadReverseCoverageContracts(t)
-	activeContracts := make([]contractDoc, 0, len(contracts))
-	for _, c := range contracts {
-		if c.Lifecycle == "active" {
-			activeContracts = append(activeContracts, c)
-		}
-	}
-
-	// Floor assertion: ≥40 total contracts (today: 47; conservative floor).
-	if len(contracts) < 40 {
-		t.Fatalf("DEAD-CONTRACT-01: floor scan failed: expected ≥40 contracts loaded, got %d (YAML scanner may be broken)", len(contracts))
-	}
-
-	// Load slice subscriber index.
-	subscribers, loadErr := loadSliceSubscribers(root)
-	if loadErr != nil {
-		t.Fatalf("DEAD-CONTRACT-01: loadSliceSubscribers: %v", loadErr)
-	}
-	subscriberIndex := make(map[string]bool) // contractID → has subscriber
-	for _, s := range subscribers {
-		subscriberIndex[s.ContractID] = true
-	}
-
-	// Load webhook receiver/dispatcher wiring index (contractID → has a
-	// webhook-receive or webhook-dispatch slice). Like subscriberIndex, this is
-	// discovered from the slice scan because contract.yaml Receivers/Dispatchers
-	// are derived (yaml:"-") and not populated by the static loader.
-	webhookWiredIndex, loadErr := loadSliceWebhookWiring(root)
-	if loadErr != nil {
-		t.Fatalf("DEAD-CONTRACT-01: loadSliceWebhookWiring: %v", loadErr)
-	}
-
-	// Build genPkgPath → contractYamlAbsPath from generated/contracts/http/iface_gen.go
-	// "// source:" comments. Needed because codegen remaps some source path segments
-	// (e.g. "internal" → "internalapi") that the dot-to-slash ID transform cannot reproduce.
-	genHTTPSourceMap, mapErr := loadGeneratedHTTPServiceMap(root, modPath)
-	if mapErr != nil {
-		t.Fatalf("DEAD-CONTRACT-01: loadGeneratedHTTPServiceMap: %v", mapErr)
-	}
-	// Invert: contractYamlAbsPath → genPkgPath (for O(1) lookup in contract loop).
-	contractPathToGenPkg := make(map[string]string, len(genHTTPSourceMap))
-	for genPkg, contractPath := range genHTTPSourceMap {
-		contractPathToGenPkg[contractPath] = genPkg
-	}
-
-	// Command dimension (#1580): codegen:true command contracts emit command_gen.go
-	// with a typed Handler interface. Map contractYamlAbsPath → generated command
-	// package path so the command branch below can require a cell implements the
-	// generated Handler (the reverse-coverage strengthening). codegen:false command
-	// contracts (e.g. the deferred dequeue/report/ack/extend-lease siblings) have NO
-	// command_gen.go, so they never appear here and skip the Handler check —
-	// only the ownerCell check applies to them.
-	genCommandSourceMap, cmdMapErr := loadGeneratedSourceMap(root, modPath, "command", "/command_gen.go")
-	if cmdMapErr != nil {
-		t.Fatalf("DEAD-CONTRACT-01: loadGeneratedSourceMap(command): %v", cmdMapErr)
-	}
-	contractPathToCmdGenPkg := make(map[string]string, len(genCommandSourceMap))
-	for genPkg, contractPath := range genCommandSourceMap {
-		contractPathToCmdGenPkg[contractPath] = genPkg
-	}
-
-	// Build the implemented-Service set by loading generated Service interfaces +
-	// cell / example impls in ONE ModeWorkspace type universe. ModeWorkspace (via
-	// LoadProductionPackages) is mandatory — NOT the Typed/ModeModule loader:
-	// examples/* are go.work satellite modules (#1556), and a GOWORK=off load
-	// silently matches only the in-root examples/demo, dropping every satellite
-	// example impl and falsely flagging their contracts as unimplemented. .All()
-	// (not .Production()) keeps the generated/contracts/http packages whose
-	// Service interfaces drive the types.Implements match. iface + impl land in
-	// one type-check universe so ImplementsInterface compares identical packages.
-	// Mirrors the loadModule funnel in archtest_test.go.
-	generatedHTTPPrefix := modPath + "/generated/contracts/http/"
-	generatedCommandPrefix := modPath + "/generated/contracts/command/"
-	cellsPrefix := modPath + "/cells/"
-	examplesPrefix := modPath + "/examples/"
-
-	type ifaceEntry struct {
-		pkgPath  string
-		ifaceTyp *types.Interface
-	}
-	type namedEntry struct {
-		named *types.Named
-	}
-
-	var genServiceIfaces []ifaceEntry
-	var genCommandHandlerIfaces []ifaceEntry
-	var cellNamedTypes []namedEntry
-
-	modules := findWorkspaceModules(t, root)
-	resolver, lpErr := typeseval.LoadProductionPackages(root, modules, false /* tests */, nil)
-	if lpErr != nil {
-		t.Fatalf("DEAD-CONTRACT-01: LoadProductionPackages: %v", lpErr)
-	}
-	// NOTE: .All() (NOT .Production()) is required here — generated/contracts/http
-	// packages carry the Service interfaces for the types.Implements match;
-	// .Production() filters generated/ and would make every http contract look
-	// unimplemented.
-	for _, pkg := range resolver.All() {
-		if pkg == nil || pkg.Types == nil {
-			continue
-		}
-		pkgPath := pkg.PkgPath
-
-		if strings.HasPrefix(pkgPath, generatedHTTPPrefix) {
-			obj := pkg.Types.Scope().Lookup("Service")
-			if obj == nil {
-				continue
-			}
-			tn, ok := obj.(*types.TypeName)
-			if !ok {
-				continue
-			}
-			named, ok := tn.Type().(*types.Named)
-			if !ok {
-				continue
-			}
-			iface, ok := named.Underlying().(*types.Interface)
-			if !ok {
-				continue
-			}
-			genServiceIfaces = append(genServiceIfaces, ifaceEntry{
-				pkgPath:  pkgPath,
-				ifaceTyp: iface.Complete(),
-			})
-			continue
-		}
-
-		// Command dimension (#1580): collect the generated Handler interface from
-		// each generated/contracts/command/* package. Same single-universe load as
-		// the HTTP Service ifaces, so types.Implements compares identical packages.
-		if strings.HasPrefix(pkgPath, generatedCommandPrefix) {
-			obj := pkg.Types.Scope().Lookup("Handler")
-			if obj == nil {
-				continue
-			}
-			tn, ok := obj.(*types.TypeName)
-			if !ok {
-				continue
-			}
-			named, ok := tn.Type().(*types.Named)
-			if !ok {
-				continue
-			}
-			iface, ok := named.Underlying().(*types.Interface)
-			if !ok {
-				continue
-			}
-			genCommandHandlerIfaces = append(genCommandHandlerIfaces, ifaceEntry{
-				pkgPath:  pkgPath,
-				ifaceTyp: iface.Complete(),
-			})
-			continue
-		}
-
-		isCells := strings.HasPrefix(pkgPath, cellsPrefix)
-		isExamples := strings.HasPrefix(pkgPath, examplesPrefix)
-		if !isCells && !isExamples {
-			continue
-		}
-
-		pkgScope := pkg.Types.Scope()
-		for _, name := range pkgScope.Names() {
-			obj := pkgScope.Lookup(name)
-			tn, ok := obj.(*types.TypeName)
-			if !ok {
-				continue
-			}
-			named, ok := tn.Type().(*types.Named)
-			if !ok {
-				continue
-			}
-			cellNamedTypes = append(cellNamedTypes, namedEntry{named: named})
-		}
-	}
-
-	// Build implemented Service set from single-pass collected types.
-	implementedPkgPaths := make(map[string]bool)
-	for _, impl := range cellNamedTypes {
-		for _, iface := range genServiceIfaces {
-			if typesutil.ImplementsInterface(impl.named, iface.ifaceTyp) {
-				implementedPkgPaths[iface.pkgPath] = true
-			}
-		}
-	}
-
-	// Command dimension (#1580): generatedCommandPkg → ≥1 cell/example type
-	// implements its Handler. Same types.Implements cross-check as Service above.
-	commandImplementedPkgPaths := make(map[string]bool)
-	for _, impl := range cellNamedTypes {
-		for _, iface := range genCommandHandlerIfaces {
-			if typesutil.ImplementsInterface(impl.named, iface.ifaceTyp) {
-				commandImplementedPkgPaths[iface.pkgPath] = true
-			}
-		}
-	}
-	// Anti-vacuity: if any codegen:true command contract exists (command_gen.go
-	// emitted → present in contractPathToCmdGenPkg), the scanner MUST have found
-	// at least one generated Handler interface. An empty iface set with non-empty
-	// source map means the Handler-lookup regressed (renamed interface, loader
-	// dropped generated/contracts/command/*) — fail loudly rather than passing the
-	// command branch vacuously.
-	if len(contractPathToCmdGenPkg) > 0 && len(genCommandHandlerIfaces) == 0 {
-		t.Fatal("DEAD-CONTRACT-01: codegen command contracts exist but no generated Handler " +
-			"interface was loaded from generated/contracts/command/* — the command Handler-impl " +
-			"cross-check would pass vacuously (#1580). Check the Handler lookup + ModeWorkspace loader.")
-	}
-
-	// Now evaluate each active contract for entry-point existence.
-	var diags []Diagnostic
-	for _, c := range activeContracts {
-		rel, relErr := filepath.Rel(root, c.FilePath)
-		if relErr != nil {
-			rel = c.FilePath
-		}
-		rel = filepath.ToSlash(rel)
-
-		switch c.Kind {
-		case "http":
-			// Look up the generated package path via the source map (correct mapping
-			// even when codegen remaps path segments like internal → internalapi).
-			genPkg := contractPathToGenPkg[c.FilePath]
-			if genPkg == "" {
-				// No generated package found for this contract: not implemented.
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active http contract " + c.ID + " has no generated package under generated/contracts/http/" +
-						" — run 'go run ./cmd/gocell generate contract <id>' or verify contract.yaml has 'codegen: true'" +
-						" (diagnostic emitted because no generated/contracts/<path>/v1/iface_gen.go was found for this contract)",
-				})
-				continue
-			}
-			if !implementedPkgPaths[genPkg] {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active http contract " + c.ID + " has no cell implementation of its generated Service interface " +
-						"(change lifecycle to draft/deprecated if not yet implemented)",
-				})
-			}
-		case "event":
-			hasPublisher := c.Endpoints.Publisher != ""
-			hasSubscriber := subscriberIndex[c.ID]
-			hasActorSubscriber := len(c.Endpoints.ActorSubscribers) > 0
-			if !hasPublisher && !hasSubscriber && !hasActorSubscriber {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active event contract " + c.ID + " has no publisher, no subscriber slice, and no actor subscriber " +
-						"(change lifecycle to draft/deprecated if unused)",
-				})
-			}
-		case "webhook":
-			// A webhook contract has impl when BOTH hold:
-			//   (1) ownerCell is declared (the explicit provider; mirrors
-			//       contractProvider() / ProviderEndpoint() post-fix semantics), and
-			//   (2) at least one slice wires it via a webhook-receive (inbound) or
-			//       webhook-dispatch (outbound) contractUsage.
-			// Receivers/Dispatchers on contract.yaml are derived (yaml:"-") and not
-			// populated by this static loader, so the wiring is discovered from the
-			// slice scan (webhookWiredIndex), exactly as event subscribers are.
-			// Currently vacuous (no active webhook contracts) but prevents a
-			// wrong-pass when PR-6 adds real webhook contracts.
-			if c.OwnerCell == "" {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active webhook contract " + c.ID + " has no ownerCell declared " +
-						"(webhook provider must be the explicit ownerCell; change lifecycle to draft/deprecated if unused)",
-				})
-			}
-			if !webhookWiredIndex[c.ID] {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active webhook contract " + c.ID + " has no webhook-receive or webhook-dispatch slice wiring " +
-						"(add a contractUsages[role=webhook-receive|webhook-dispatch] entry to a slice, " +
-						"or change lifecycle to draft/deprecated if unused)",
-				})
-			}
-		case "command":
-			// Every command contract needs a provider: ownerCell or the command's
-			// provider endpoint (endpoints.handler, per metadata ProviderEndpoint —
-			// covers codegen:false siblings too). Using contractProviderEndpoint
-			// rather than endpoints.server avoids false-flagging a kind:command that
-			// declares only endpoints.handler (ownerCell deriving from handler).
-			if c.OwnerCell == "" && contractProviderEndpoint(c) == "" {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active command contract " + c.ID + " has no ownerCell or " +
-						contractProviderFieldLabel(c.Kind) + " declared " +
-						"(change lifecycle to draft/deprecated if unused)",
-				})
-			}
-			// codegen:true command contracts additionally require a cell that
-			// implements the generated Handler interface — mirroring HANDLER-DECL-COVER-01
-			// for HTTP Service (#1580). codegen:false commands have no generated
-			// package (absent from contractPathToCmdGenPkg) and skip this check.
-			if genPkg := contractPathToCmdGenPkg[c.FilePath]; genPkg != "" && !commandImplementedPkgPaths[genPkg] {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active codegen command contract " + c.ID + " has no cell implementation of its " +
-						"generated Handler interface (implement the generated <pkg>.Handler in a cell/example slice " +
-						"and register it via <pkg>.Register, or change lifecycle to draft/deprecated / codegen: false " +
-						"if not yet implemented)",
-				})
-			}
-		default:
-			// projection / grpc / saga / any other kind: require ownerCell or the
-			// kind-correct provider endpoint (projection→endpoints.provider,
-			// grpc/saga→endpoints.server), read via contractProviderEndpoint so the
-			// per-kind provider field matches metadata ProviderEndpoint.
-			if c.OwnerCell == "" && contractProviderEndpoint(c) == "" {
-				diags = append(diags, Diagnostic{
-					Rel:  rel,
-					Line: 1,
-					Message: "active " + c.Kind + " contract " + c.ID + " has no ownerCell or " +
-						contractProviderFieldLabel(c.Kind) + " declared " +
-						"(change lifecycle to draft/deprecated if unused)",
-				})
-			}
-		}
-	}
-	Report(t, "DEAD-CONTRACT-01", diags)
+	Report(t, "DEAD-CONTRACT-01", CheckDeadContractCover(t, ConfigForExternalCell{}))
 }
 
 // TestDeadContractCover_FloorScan is the dedicated floor-scan self-check:
@@ -1487,7 +427,7 @@ func TestDeadContractCover_FloorScan(t *testing.T) {
 // pass vacuously and never surface an unimplemented codegen command.
 func TestDeadContractCover_DetectsUnimplementedCommand(t *testing.T) {
 	t.Parallel()
-	const fakeCmdGenPkg = "github.com/ghbvf/gocell/generated/contracts/command/fake/op/v1"
+	const fakeCmdGenPkg = PlatformModulePath + "/generated/contracts/command/fake/op/v1"
 	const fakeContractPath = "/abs/examples/x/contracts/command/fake/op/v1/contract.yaml"
 
 	// Simulate: this codegen command contract maps to a generated package, but no
@@ -1593,102 +533,7 @@ func TestDeadContractCover_ProviderEndpointMirrorsMetadata(t *testing.T) {
 // way as an http one. Handles "internal" → "internalapi" codegen remapping.
 func TestDeadCodeCover(t *testing.T) {
 	t.Parallel()
-	root := findModuleRoot(t)
-	modPath, err := moduleImportPath(root)
-	if err != nil {
-		t.Fatalf("DEAD-CODE-01: read module path: %v", err)
-	}
-
-	contracts := loadReverseCoverageContracts(t)
-	var deprecated []contractDoc
-	for _, c := range contracts {
-		if c.Lifecycle == "deprecated" {
-			deprecated = append(deprecated, c)
-		}
-	}
-
-	// Vacuous pass when no deprecated contracts exist.
-	if len(deprecated) == 0 {
-		t.Logf("DEAD-CODE-01: 0 deprecated contracts (vacuous pass); rule will activate when first deprecated contract is added")
-		return
-	}
-
-	// Build deprecated gen-pkg set via the all-kinds source map (covers
-	// http/event/projection; correctly handles "internal" → "internalapi").
-	genAllSourceMap, mapErr := loadGeneratedAllSourceMap(root, modPath)
-	if mapErr != nil {
-		t.Fatalf("DEAD-CODE-01: loadGeneratedAllSourceMap: %v", mapErr)
-	}
-	// Build reverse: contractYamlAbsPath → genPkgPath.
-	contractPathToGenPkg := make(map[string]string, len(genAllSourceMap))
-	for genPkg, contractPath := range genAllSourceMap {
-		contractPathToGenPkg[contractPath] = genPkg
-	}
-
-	// Build deny sets.
-	deprecatedIDs := make(map[string]bool)
-	deprecatedGenPkgs := make(map[string]bool)
-	for _, c := range deprecated {
-		deprecatedIDs[c.ID] = true
-		// Use source map for http contracts (handles internal → internalapi).
-		if genPkg, ok := contractPathToGenPkg[c.FilePath]; ok {
-			deprecatedGenPkgs[genPkg] = true
-		}
-		// For non-http contracts (event/command/etc.) that have no generated
-		// package in generated/contracts/, the source map entry is absent and
-		// we do not need to add a gen-pkg entry (no import to scan for).
-	}
-
-	scope := ModuleScope(root, MatchRels(func(rel string) bool {
-		return strings.HasSuffix(rel, ".go") &&
-			!strings.HasSuffix(rel, "_test.go") &&
-			!strings.Contains(rel, "/testdata/")
-	}))
-
-	diags := Run(t, AST(scope), func(p *Pass) []Diagnostic {
-		var d []Diagnostic
-		for _, f := range p.Files {
-			rel := p.Rel(f)
-
-			for _, imp := range f.Imports {
-				if imp.Path == nil {
-					continue
-				}
-				impPath := strings.Trim(imp.Path.Value, `"`)
-				if deprecatedGenPkgs[impPath] {
-					pos := p.Fset.Position(imp.Pos())
-					d = append(d, Diagnostic{
-						Rel:  rel,
-						Line: pos.Line,
-						Message: "imports deprecated contract package " + impPath +
-							" (remove all references to deprecated contracts; lifecycle must be updated or code deleted)",
-					})
-				}
-			}
-
-			EachInSubtree[ast.BasicLit](f, func(lit *ast.BasicLit) {
-				if lit.Kind.String() != "STRING" {
-					return
-				}
-				val, ok := StringLitValue(lit)
-				if !ok {
-					return
-				}
-				if deprecatedIDs[val] {
-					pos := p.Fset.Position(lit.Pos())
-					d = append(d, Diagnostic{
-						Rel:  rel,
-						Line: pos.Line,
-						Message: "references deprecated contract ID " + val +
-							" as string literal (remove all references to deprecated contracts)",
-					})
-				}
-			})
-		}
-		return d
-	})
-
-	Report(t, "DEAD-CODE-01", diags)
+	Report(t, "DEAD-CODE-01", CheckDeadCodeCover(t, ConfigForExternalCell{}))
 }
 
 // TestDeadCodeCover_DetectsDeprecatedImport is the negative self-check for
@@ -1707,7 +552,7 @@ func TestDeadCodeCover(t *testing.T) {
 func TestDeadCodeCover_DetectsDeprecatedImport(t *testing.T) {
 	t.Parallel()
 
-	const fakeGenPkg = "github.com/ghbvf/gocell/generated/contracts/event/deprecated/v1"
+	const fakeGenPkg = PlatformModulePath + "/generated/contracts/event/deprecated/v1"
 	const fakeContractID = "event.deprecated.v1"
 
 	deprecatedGenPkgs := map[string]bool{fakeGenPkg: true}
